Change `char port` to `int port`
The C++ functions that accept `port` as an arugment expect an integer value. Some of the Python wrapping functions declare `char port` rather than `int port` which results in unexected behaviour. e.g. passing `1` as a string to the function will actually try and open port `0x31` which is the hex representation of `1`.

The `Py_ndiPVWRFromFile` function was correctly using `int port`, but all other uses of `port` were of the incorrect type.